[Decomp][Unit] Cohesion split: extract diminishing-returns tracking to UnitDiminishing.cpp
