+++ src/game/Object/Unit.cpp
@@ -6146,153 +6146,6 @@ int32 Unit::CalculateAuraDuration(SpellEntry const* spellProto, uint32 effectMas
 }
 
 /**
- * @brief Gets the current diminishing return level for a control group.
- *
- * @param group The diminishing return group.
- * @return The current diminishing level.
- */
-DiminishingLevels Unit::GetDiminishing(DiminishingGroup group)
-{
-    for (Diminishing::iterator i = m_Diminishing.begin(); i != m_Diminishing.end(); ++i)
-    {
-        if (i->DRGroup != group)
-        {
-            continue;
-        }
-
-        if (!i->hitCount)
-        {
-            return DIMINISHING_LEVEL_1;
-        }
-
-        if (!i->hitTime)
-        {
-            return DIMINISHING_LEVEL_1;
-        }
-
-        // If last spell was casted more than 15 seconds ago - reset the count.
-        if (i->stack == 0 && getMSTimeDiff(i->hitTime, GameTime::GetGameTimeMS()) > 15 * IN_MILLISECONDS)
-        {
-            i->hitCount = DIMINISHING_LEVEL_1;
-            return DIMINISHING_LEVEL_1;
-        }
-        // or else increase the count.
-        else
-        {
-            return DiminishingLevels(i->hitCount);
-        }
-    }
-    return DIMINISHING_LEVEL_1;
-}
-
-/**
- * @brief Increments the diminishing return counter for a control group.
- *
- * @param group The diminishing return group.
- */
-void Unit::IncrDiminishing(DiminishingGroup group)
-{
-    // Checking for existing in the table
-    for (Diminishing::iterator i = m_Diminishing.begin(); i != m_Diminishing.end(); ++i)
-    {
-        if (i->DRGroup != group)
-        {
-            continue;
-        }
-        if (i->hitCount < DIMINISHING_LEVEL_IMMUNE)
-        {
-            i->hitCount += 1;
-        }
-        return;
-    }
-    m_Diminishing.push_back(DiminishingReturn(group, GameTime::GetGameTimeMS(), DIMINISHING_LEVEL_2));
-}
-
-/**
- * @brief Applies diminishing return rules to a control effect duration.
- *
- * @param group The diminishing return group.
- * @param duration The mutable duration in milliseconds.
- * @param caster The caster applying the effect.
- * @param Level The diminishing level being applied.
- * @param isReflected True when the effect was reflected.
- */
-void Unit::ApplyDiminishingToDuration(DiminishingGroup group, int32& duration, Unit* caster, DiminishingLevels Level, int32 limitduration, bool isReflected)
-{
-    if (duration == -1 || group == DIMINISHING_NONE || (!isReflected && caster->IsFriendlyTo(this)))
-    {
-        return;
-    }
-
-    // Duration of crowd control abilities on pvp target is limited by 10 sec. (2.2.0)
-    if (limitduration > 0 && duration > limitduration)
-    {
-        // test pet/charm masters instead pets/charmeds
-        Unit const* targetOwner = GetCharmerOrOwner();
-        Unit const* casterOwner = caster->GetCharmerOrOwner();
-
-        Unit const* target = targetOwner ? targetOwner : this;
-        Unit const* source = casterOwner ? casterOwner : caster;
-
-        if (target->GetTypeId() == TYPEID_PLAYER && source->GetTypeId() == TYPEID_PLAYER)
-        {
-            duration = limitduration;
-        }
-    }
-
-    float mod = 1.0f;
-
-    // Some diminishings applies to mobs too (for example, Stun)
-    if ((GetDiminishingReturnsGroupType(group) == DRTYPE_PLAYER && GetTypeId() == TYPEID_PLAYER) || GetDiminishingReturnsGroupType(group) == DRTYPE_ALL)
-    {
-        DiminishingLevels diminish = Level;
-        switch (diminish)
-        {
-            case DIMINISHING_LEVEL_1: break;
-            case DIMINISHING_LEVEL_2: mod = 0.5f; break;
-            case DIMINISHING_LEVEL_3: mod = 0.25f; break;
-            case DIMINISHING_LEVEL_IMMUNE: mod = 0.0f; break;
-            default: break;
-        }
-    }
-
-    duration = int32(duration * mod);
-}
-
-/**
- * @brief Updates active stack tracking for a diminishing return group.
- *
- * @param group The diminishing return group.
- * @param apply True when an aura from the group is applied; false when it is removed.
- */
-void Unit::ApplyDiminishingAura(DiminishingGroup group, bool apply)
-{
-    // Checking for existing in the table
-    for (Diminishing::iterator i = m_Diminishing.begin(); i != m_Diminishing.end(); ++i)
-    {
-        if (i->DRGroup != group)
-        {
-            continue;
-        }
-
-        if (apply)
-        {
-            i->stack += 1;
-        }
-        else if (i->stack)
-        {
-            i->stack -= 1;
-            // Remember time after last aura from group removed
-            if (i->stack == 0)
-            {
-                i->hitTime = GameTime::GetGameTimeMS();
-            }
-        }
-        break;
-    }
-}
-
-/**
  * @brief Checks whether the unit is visible to a player without stealth detection.
  *
  * @param u The observing player.
